refacto(Tooltip): extract tooltip logic to component
