+++ src/components/slides/TemperatureSlide/Tooltip/Tooltip.css
@@ -1,0 +1,12 @@
+.tooltip{
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    background-color: white ;
+    border: solid ;
+    border-width:1px;
+    border-radius: 5px;
+    padding: 10px;
+    position: relative;
+    visibility: hidden;
+}
