+++ src/components/slides/TemperatureSlide/Tooltip/Tooltip.jsx
@@ -1,0 +1,10 @@
+import React from "react";
+import "./Tooltip.css";
+
+
+export const Tooltip = ( ) => {
+
+    return (
+        <div className="tooltip">{}</div>
+    );
+};
